models/requirements.txt
flask 
pillow 
sentence-transformers
torch